Created an extended project markdown project
The extended project adds an additional phase of processing where the
train/test data are scaled before building the set of candidate ML
models.  The same statistical analysis is then done on the
cross-validation results and then compared with those using the
non-scaled train/test datasets.
This is building on top of the base project that was submitted as part
of the Coursera Practical Machine Learning Course.